roomWeatherStation/roomWeatherStation/requirements.txt
APScheduler==3.6.3
click==7.1.2
Flask==1.1.2
Flask-SQLAlchemy==2.4.3
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
pyserial==3.4
pytz==2020.1
six==1.15.0
SQLAlchemy==1.3.18
tzlocal==2.1
Werkzeug==1.0.1
